+++ src/Enum/LastClause.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FQL\Enum;
+
+enum LastClause: string
+{
+    case SELECT = 'select';
+    case FROM = 'from';
+    case JOIN = 'join';
+}